Refactoring metrics_helper
